v7.0

+++ src/main/java/com/foxminded/StudentMarks.java
@@ -10,7 +10,7 @@ public class StudentMarks {
     private long id;
     private StudentCard studentCard;
     private long sectionId;
-    private List<Integer> marks = new ArrayList<>();
+    private List<Byte> marks = new ArrayList<>();
 
     public StudentMarks(){}
 
@@ -23,14 +23,14 @@ public class StudentMarks {
         if (marks.size() == 0) return 0;
         double result =  0.0;
         int counter = 0;
-        for(Integer mark: marks){
+        for(Byte mark: marks){
             result += mark;
             counter++;
         }
         return result/counter;
     }
 
-    public void addMark(int mark) throws ValidationException{
+    public void addMark(byte mark) throws ValidationException{
         if ((mark < 1) || (mark > 12)) throw new ValidationException("Mark should be [1;12]");
         marks.add(mark);
     }
